Remove navigator and added folding
