+++ plugin/treesitter_config.lua
@@ -1,4 +1,7 @@
-require "nvim-treesitter.configs".setup {
+local configs = require "nvim-treesitter.configs";
+
+configs.setup {
+	ensure_installed = "maintained",
     highlight = {
         enable = true,
         custom_captures = {},
